refactor: seperate answer api (#102)
