+++ src/api/index.ts
@@ -1,3 +1,4 @@
+export * from './answer'
 export * from './api'
 export * from './chat-api'
 export * from './qna-api'
